+++ druid/examples/hello.rs
@@ -40,6 +40,7 @@ pub fn main() {
 
     // start the application
     AppLauncher::with_window(main_window)
+        .use_simple_logger()
         .launch(initial_state)
         .expect("Failed to launch application");
 }